// lib/interfaces/csc.interface.ts
export interface ICountry {
    id: number;
    name: string;
    iso3: string;
    iso2: string;
    phone_code: string;
    capital: string;
    currency: string;
}

export interface IState {
    id: number;
    name: string;
    country_id: number;
    country_code: string;
    state_code: string;
}

export interface ICity {
    id: number;
    name: string;
    state_id: number;
    state_code: string;
    country_id: number;
    country_code: string;
    latitude: string;
    longitude: string;
}